Move View+SizeTracker to WooFoundation
