+++ WooCommerce/Classes/ViewRelated/Orders/Order Creation/CustomAmounts/AddCustomAmountPercentageView.swift
@@ -1,5 +1,4 @@
 import SwiftUI
-import struct WooFoundation.AutofocusTextModifier
 
 struct AddCustomAmountPercentageView: View {
     @ObservedObject private(set) var viewModel: AddCustomAmountPercentageViewModel
