+++ WooCommerce/Classes/ViewRelated/Orders/Order Creation/PaymentSection/GiftCardInputView.swift
@@ -1,6 +1,5 @@
 import SwiftUI
 import struct WooFoundation.WooRoundedBorderTextFieldStyle
-import struct WooFoundation.AutofocusTextModifier
 
 /// Allows the user to enter a gift card code.
 struct GiftCardInputView: View {
