+++ WooCommerce/Classes/ViewRelated/Orders/Order Creation/ProductsSection/DiscountLineDetailsView.swift
@@ -1,5 +1,4 @@
 import SwiftUI
-import struct WooFoundation.AutofocusTextModifier
 
 struct DiscountLineDetailsView: View {
     @ObservedObject private var viewModel: FeeOrDiscountLineDetailsViewModel
