+++ WooCommerce/Classes/ViewRelated/Orders/Order Details/Order Notes Section/Customer Note/EditCustomerNote.swift
@@ -1,7 +1,6 @@
 import Foundation
 import Combine
 import SwiftUI
-import struct WooFoundation.AutofocusTextModifier
 
 /// Hosting controller that wraps an `EditCustomerNote` view.
 ///
