+++ WooCommerce/Classes/ViewRelated/Reviews/ReviewReply.swift
@@ -1,7 +1,6 @@
 import Foundation
 import SwiftUI
 import Combine
-import struct WooFoundation.AutofocusTextModifier
 
 /// Hosting controller that wraps a `ReviewDetailsReply` view.
 ///
